# accumulator/utils: Much much faster treeRows