--- a/accumulator/utils.go
+++ b/accumulator/utils.go
@@ -2,6 +2,7 @@ package accumulator
 
 import (
 	"fmt"
+	"math"
 	"sort"
 )
 
@@ -181,12 +182,33 @@ func inForest(pos, numLeaves uint64, forestRows uint8) bool {
 	return pos < numLeaves
 }
 
-// given n leaves, how deep is the tree?
-// iterate shifting left until greater than n
-func treeRows(n uint64) (e uint8) {
-	for ; (1 << e) < n; e++ {
-	}
-	return
+// treeRows returns the number of rows given n leaves.
+func treeRows(n uint64) uint8 {
+	// treeRows works by:
+	// 1. Find the next power of 2 from the given n leaves.
+	// 2. Calculate the log2 of the result from step 1.
+	//
+	// For example, if the given number is 9, the next power of 2 is
+	// 16. This log2 of this number is how many rows there are in the
+	// given tree.
+	//
+	// This works because while Utreexo is a collection of perfect
+	// trees, the allocated number of leaves is always a power of 2.
+	// For Utreexo trees that don't have leaves that are power of 2,
+	// the extra space is just unallocated/filled with zeros.
+
+	// Find the next power of 2
+	n--
+	n |= n >> 1
+	n |= n >> 2
+	n |= n >> 4
+	n |= n >> 8
+	n |= n >> 16
+	n |= n >> 32
+	n++
+
+	// log of 2 is the tree depth/height
+	return uint8(math.Log2(float64(n)))
 }
 
 // numRoots is just a popCount function, returning the number of 1 bits