```treeRows``` is called fairly often and from pprof results, treeRows is fairly high in the list of time taken. The current implementation is just bit shifting. I've implemented as:

1. Find the next power of 2
2. Take the log2 of that number

From this, I've gone ahead and implemented some bit tricks to really optimize it. 

![Screen Shot 2020-08-06 at 2 02 31 AM](https://user-images.githubusercontent.com/37185887/89443150-dce9f400-d78a-11ea-92be-1731af657161.png)

With the new implementation (thanks @dergoegge) treeRows now only takes ~2% of the time it originally took. My benchmarks are done on a laptop so it's not as good but the results from benchstat are:

```
[I] calvin@bitcoin ~/b/u/g/s/g/m/u/accumulator> benchstat old.txt new.txt
name                    old time/op  new time/op  delta
OrigTreeRows_HunThou-4  0.00ns ±19%  0.00ns ±13%   -98.51%  (p=0.008 n=5+5)
OrigTreeRows_Mil-4      0.02ns ± 1%  0.00ns ±13%   -97.71%  (p=0.008 n=5+5)
OrigTreeRows_Bil-4      0.22ns ± 1%  0.01ns ±35%   -97.47%  (p=0.008 n=5+5)
OrigTreeRows_Tril-4      2.50s ± 1%   0.00s ±12%  -100.00%  (p=0.008 n=5+5)
```


The savings increase as the number of times rem2 is called.

The actual bit trick algorithms implemented are modified from https://graphics.stanford.edu/~seander/bithacks.html#DetermineIfPowerOf2