@adiabat 2020-08-07
Cool!  I don't really know how this works (De Bruijn sequence...) but does seem to work, and faster.

(also, nitpick but the benchmerk functions are called "Mi Bil Tril" but are not million / billion / trillion so... eh whatever :)